- Fix build on 64 bit architectures (cast from pointer to int loses precision) - Drop @dirrm* from plist - Add empty directories to plist
Approved by:	portmgr blanket

+++ games/gtkradiant/files/patch-radiant_treemodel.cpp
@@ -9,7 +9,7 @@
 +  // C++ code, and Clang will assume that comparison below always evaluates
 +  // to true, resulting in a segmentation fault.  Use a dirty hack to force
 +  // Clang to check those "bad" references for null nonetheless.
-+  volatile int n = (int)&node;
++  volatile intptr_t n = (intptr_t)&node;
 +
 +  if(n != 0)
    {
@@ -20,7 +20,7 @@
  void node_detach_name_changed_callback(scene::Node& node, const NameCallback& callback)
  {
 -  if(&node != 0)
-+  volatile int n = (int)&node;		// see the comment on line 713
++  volatile intptr_t n = (intptr_t)&node;		// see the comment on line 713
 +
 +  if(n != 0)
    {
@@ -31,7 +31,7 @@
  const char* node_get_name_safe(scene::Node& node)
  {
 -  if(&node == 0)
-+  volatile int n = (int)&node;		// see the comment on line 713
++  volatile intptr_t n = (intptr_t)&node;		// see the comment on line 713
 +
 +  if(n == 0)
    {
@@ -42,7 +42,7 @@
  void node_attach_name_changed_callback(scene::Node& node, const NameCallback& callback)
  {
 -  if(&node != 0)
-+  volatile int n = (int)&node;		// see the comment on line 713
++  volatile intptr_t n = (intptr_t)&node;		// see the comment on line 713
 +
 +  if(n != 0)
    {
@@ -53,7 +53,7 @@
  void node_detach_name_changed_callback(scene::Node& node, const NameCallback& callback)
  {
 -  if(&node != 0)
-+  volatile int n = (int)&node;		// see the comment on line 713
++  volatile intptr_t n = (intptr_t)&node;		// see the comment on line 713
 +
 +  if(n != 0)
    {
